transactions everything works fine budgets not

+++ app/src/main/java/de/christcoding/budgetfellow/viewmodels/AddOrEditTransactionViewModel.kt
@@ -52,6 +52,7 @@ open class AddOrEditTransactionViewModel(
     var stepDesc by mutableStateOf("Let's start by adding your first fix income.")
     var transactionName by mutableStateOf("")
     var transactionId = 0L
+    var recurringId = ""
     var transactionDescription by mutableStateOf("")
     var amount by mutableStateOf("")
     var id by mutableStateOf(-2L)
@@ -111,7 +112,8 @@ open class AddOrEditTransactionViewModel(
             }
             is AddTransactionEvent.OnDeleteRecurringClicked -> {
                     for (transaction in transactions) {
-                        if (transaction.recurringId.equals(_editableTransaction!!.recurringId)) {
+                        if (transaction.recurringId.equals(_editableTransaction?.recurringId)
+                            && transaction.date.isAfter(_editableTransaction?.date?.minusDays(1))) {
                             viewModelScope.launch {
                                 transactionRepository.deleteATransaction(transaction)
                         }
@@ -204,7 +206,8 @@ open class AddOrEditTransactionViewModel(
                 date = datePicked,
                 recurring = recurring,
                 recurringInterval = recurringPeriod.toInt(),
-                recurringIntervalUnit = periodUnit
+                recurringIntervalUnit = periodUnit,
+                recurringId = UUID.randomUUID().toString()
             )
         if (transactionId > 0) {
             ta = ta.copy(id = transactionId)
@@ -267,6 +270,7 @@ open class AddOrEditTransactionViewModel(
             }
             _editableTransaction = trans
             transactionId = trans.id
+            recurringId = trans.recurringId
             transactionName = trans.name
             transactionDescription = trans.description
             amount = trans.amount.toString()
@@ -281,7 +285,6 @@ open class AddOrEditTransactionViewModel(
                 period = trans.recurringInterval.toString()
             )
         }
-
     }
 
     private fun isInvalid(trans: TransactionDetails): Boolean {
